first pass at updating to work with ESM and split into 3 project types
