+++ npm-dependency/index.html
@@ -1,0 +1,33 @@
+<!--
+ Open MCT, Copyright (c) 2014-2020, United States Government
+ as represented by the Administrator of the National Aeronautics and Space
+ Administration. All rights reserved.
+ Open MCT is licensed under the Apache License, Version 2.0 (the
+ "License"); you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+ http://www.apache.org/licenses/LICENSE-2.0.
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ License for the specific language governing permissions and limitations
+ under the License.
+ Open MCT includes source code licensed under additional open source
+ licenses. See the Open Source Licenses file (LICENSES.md) included with
+ this source code distribution or the Licensing information page available
+ at runtime from the About dialog for additional information.
+-->
+
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Open MCT as a Dependency Example</title>
+    <script src="/node_modules/openmct/dist/openmct.js"></script>
+    <script src="/openmct-yamcs-example.js" defer></script>
+
+    <link rel="icon" type="image/png" href="/node_modules/openmct/dist/favicons/favicon-96x96.png" sizes="96x96" type="image/x-icon">
+    <link rel="icon" type="image/png" href="/node_modules/openmct/dist/favicons/favicon-32x32.png" sizes="32x32" type="image/x-icon">
+    <link rel="icon" type="image/png" href="/node_modules/openmct/dist/favicons/favicon-16x16.png" sizes="16x16" type="image/x-icon">
+</head>
+<body>
+</body>
+</html>
