+++ npm-dependency/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "openmct-as-a-npm-dependency",
+  "version": "1.0.0",
+  "description": "Open MCT As A dependency",
+  "private": true,
+  "scripts": {
+    "clean": "rm -rf ./dist ./node_modules ./package-lock.json ./openmct/node_modules ./openmct/package-lock.json",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack",
+    "start": "webpack serve"
+  },
+  "author": "National Aeronautics and Space Administration",
+  "license": "ISC",
+  "dependencies": {},
+  "devDependencies": {
+    "@babel/core": "^7.13.10",
+    "babel-loader": "9.1.0",
+    "copy-webpack-plugin": "12.0.2",
+    "css-loader": "6.10.0",
+    "html-loader": "^0.5.5",
+    "http-server": "^0.12.3",
+    "mini-css-extract-plugin": "2.7.6",
+    "sass": "1.71.1",
+    "sass-loader": "14.1.1",
+    "openmct": "nasa/openmct#master",
+    "vue": "3.4.19",
+    "vue-loader": "16.8.3",
+    "vue-template-compiler": "^2.6.12",
+    "webpack": "5.90.3",
+    "webpack-cli": "5.1.1",
+    "webpack-dev-server": "5.0.2"
+  }
+}
