--- src/plugins/HelloWorld/components/HelloWorldComponent.vue
@@ -1,11 +1,0 @@
-<template>
-<div>
-    {{ domainObject.string }} For Hello World
-</div>
-</template>
-
-<script>
-export default {
-    inject: ['openmct', 'domainObject']
-};
-</script>
